+++ Services/Commands/DeleteRoom/DeleteRoomCommandHandler.cs
@@ -1,4 +1,5 @@
-﻿using Domain.Repositories;
+﻿using Application.Common.Exceptions;
+using Domain.Repositories;
 using MediatR;
 
 namespace Application.Commands.DeleteRoom;
@@ -15,6 +16,10 @@ internal class DeleteRoomCommandHandler : IRequestHandler<DeleteRoomCommand>
     public async Task<Unit> Handle(DeleteRoomCommand request, CancellationToken cancellationToken)
     {
         var room = await _roomRepository.GetRoomByIdAsync(request.Id);
+        if (room == null)
+        {
+            throw new NotFoundException(nameof(room), request.Id);
+        }
 
         await _roomRepository.DeleteRoomAsync(room);
